server/requirements.txt
alembic==1.7.5
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
flask-socketio
python-dotenv==0.20.0
greenlet
importlib-metadata==4.9.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mpmath==1.2.1
paramiko==2.9.0
psycopg2-binary==2.9.2
pycparser==2.21
PyNaCl==1.4.0
six==1.16.0
simple-websocket
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
marshmallow
